Introduce AbstractColumnMoveOperation
